Added Namespaces
Folder structure changed
