+++ Clash Royal/Assets/Scripts/Chest/ChestSystem/Controller.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5d1308e7c442340449ed93386f3e1af4
+guid: 632952960d17a6347a6db4d8d9cbff46
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
