+++ Clash Royal/Assets/Scripts/Chest/ChestSystem/Model.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1747bc47598e7454ca8d0a800a498f3a
+guid: 24133d58356dc134da2a4d174e24dbf5
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
