+++ Clash Royal/Assets/Scripts/Chest/ChestSystem/Service Locator/ServiceLocatorInstaller.cs
@@ -1,6 +1,5 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
+using ChestSystem.Utility;
 
 public class ServiceLocatorInstaller : MonoBehaviour
 {
