+++ Clash Royal/Assets/Scripts/Chest/ChestSystem/State Machine.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1c8ba9de7b4c0394ebcc7129be5fd4c0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
