+++ Clash Royal/Assets/Scripts/Chest/ChestSystem/Utility.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e80e78ef1e2a9c647ad357decf657940
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
