+++ Clash Royal/Assets/Scripts/Chest/ChestSystem/Utility/IChestObserver.cs
@@ -1,0 +1,12 @@
+using ChestSystem.Model;
+
+
+namespace ChestSystem.Utility
+{
+    //Observer Pattern
+    public interface IChestObserver
+    {
+        void OnChestSpawned(ChestModel chestModel, int slotIndex);
+    }
+}
+
