+++ Clash Royal/Assets/Scripts/Chest/ChestSystem/Utility/ICommand.cs
@@ -1,0 +1,12 @@
+//Command Pattern to execute and undo
+
+namespace ChestSystem.Utility
+{
+    public interface ICommand
+    {
+        void Execute();
+        void Undo();
+    }
+}
+
+
