+++ Clash Royal/Assets/Scripts/Chest/ChestSystem/View.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6e5b78efdd9f6334b9a2a4f641b46937
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
