--- Clash Royal/Assets/Scripts/Chest/Interface/ICommand.cs
@@ -1,6 +1,0 @@
-//Command Pattern to execute and undo
-public interface ICommand
-{
-    void Execute();
-    void Undo();
-}
