--- Clash Royal/Assets/Scripts/Chest/Observer/IChestObserver.cs
@@ -1,5 +1,0 @@
-//Observer Pattern
-public interface IChestObserver
-{
-   void OnChestSpawned(ChestModel chestModel, int slotIndex);
-}
